+++ src/address.h
@@ -135,7 +135,7 @@ public:
     //返回端口号
     virtual uint32_t getPort() const = 0;
     //设置端口号
-    virtual void setPort(uint32_t v) = 0;
+    virtual void setPort(uint16_t v) = 0;
 };
 
 class IPv4Address : public IPAddress{
@@ -167,7 +167,7 @@ public:
     IPAddress::ptr subnetMask(uint32_t prefix_len) override;
 
     uint32_t getPort() const override;
-    void setPort(uint32_t v) override;
+    void setPort(uint16_t v) override;
 private:
     sockaddr_in m_addr;
 };
@@ -200,7 +200,7 @@ public:
     IPAddress::ptr subnetMask(uint32_t prefix_len) override;
 
     uint32_t getPort() const override;
-    void setPort(uint32_t v) override;
+    void setPort(uint16_t v) override;
 private:
     sockaddr_in6 m_addr;
 };
